pyproject.toml
[project]
name = "tfds"
version = "0.1.1"
description = "The free data stack CLI"
authors = [
    {name = "Jens Köster",email = "tfds@jensbox.se"}
]
license = {text = "GNU GENERAL PUBLIC LICENSE v3"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "typer[all]>=0.10.0",
    "boto3 (>=1.38.19,<2.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "nbformat (>=5.10.4,<6.0.0)",
    "tfdslib (>=0.1.4,<0.2.0)",
    "gitpython (>=3.1.44,<4.0.0)",
]

[project.scripts]
tfds = "tfds_cli.cli:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [
  { include = "tfds_cli", from = "src" }
]

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"
mypy = "^1.15.0"
pytest = "^8.3.5"
pre-commit = "^4.2.0"
ruff = "^0.11.10"



[[tool.poetry.source]]
name = "devpi"
url = "http://devpi:8008/root/tfds/+simple/"
priority = "supplemental"

[tool.black]
line-length = 88

[tool.isort]
profile = "black"

[tool.mypy]
python_version = "3.11"
strict = true


[[tool.mypy.overrides]]
module = "tests.*"
disable_error_code = ["no-untyped-def"]

[[tool.mypy.overrides]]
module = "yaml"
ignore_missing_imports = true

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
ignore-init-module-imports = true

[tool.ruff.lint.per-file-ignores]
"tests/*.py" = ["ANN201"]